refactor: consolidate CSL auto-config in WebSecurityConfig and rename slice
With the pt-security profile gone and consolidated-auth always active when
security wiring is under test, the @ImportAutoConfiguration belongs alongside
the rest of the host's security wiring on WebSecurityConfig. Revert the
CamundaSecurityLibraryImporter extraction and move the SecurityPathPort and
WebAppProviderPort bean methods back to WebSecurityConfig too.

Rename TenantSecuritySlice to PhysicalTenantChainContext for a clearer name
that describes the role of the per-tenant object the chain factory consumes.
